[fix] address various bugs and restructure models
